Remove Servant.:> from the Prelude
