+++ src/Monocle/Servant/HTTP.hs
@@ -5,7 +5,7 @@
 -- |
 -- Copyright: (c) 2021 Monocle authors
 -- SPDX-License-Identifier: AGPL-3.0-only
-module Monocle.Servant.HTTP (MonocleAPI, server) where
+module Monocle.Servant.HTTP (RootAPI, MonocleAPI, server) where
 
 import Monocle.Api.Jwt (AuthenticatedUser)
 import Monocle.Api.Server (authGetMagicJwt, authWhoAmi, configGetAbout, configGetGroupMembers, configGetGroups, configGetProjects, configGetWorkspaces, crawlerAddDoc, crawlerCommit, crawlerCommitInfo, loginLoginValidation, metricGet, metricInfo, metricList, searchAuthor, searchCheck, searchFields, searchQuery, searchSuggestions)
@@ -19,12 +19,15 @@ import Monocle.Protob.Search (AuthorRequest, AuthorResponse, CheckRequest, Check
 import Monocle.Servant.PBJSON (PBJSON)
 import Servant
 import Servant.Auth.Server (Auth, Cookie, JWT)
+import Data.Text
 
 import Effectful (Eff)
 import Monocle.Effects hiding (searchQuery)
 import Effectful.Concurrent.MVar qualified as E
 import Effectful ((:>>))
 import Prelude (undefined)
+import Monocle.Api.Jwt (LoginInUser (..))
+import Servant.HTML.Blaze (HTML)
 
 type MonocleAPI =
   "login" :> "username" :> "validate" :> Auth '[JWT, Cookie] AuthenticatedUser :> ReqBody '[JSON] Monocle.Protob.Login.LoginValidationRequest :> Post '[PBJSON, JSON] Monocle.Protob.Login.LoginValidationResponse
@@ -67,3 +70,13 @@ server =
     :<|> crawlerAddDoc
     :<|> crawlerCommit
     :<|> crawlerCommitInfo
+
+-- | The API is served at both `/api/2/` (for backward compat with the legacy nginx proxy)
+-- and `/` (for compat with crawler client)
+type MonocleAPI' = MonocleAPI :<|> AuthAPI
+
+type RootAPI = "api" :> "2" :> MonocleAPI' :<|> MonocleAPI'
+
+type AuthAPI =
+  "auth" :> "login" :> QueryParam "redirectUri" Text :> Get '[JSON] NoContent
+    :<|> "auth" :> "cb" :> QueryParam "error" Text :> QueryParam "code" Text :> QueryParam "state" Text :> Get '[HTML] LoginInUser
